demo45/requirements.txt
numpy
nltk
faiss-cpu
sentence-transformers
google-generativeai
python-dotenv
pdfplumber
sqlalchemy
